+++ finagle-base-http/src/main/scala/com/twitter/finagle/http/exp/GenStreamingSerialServerDispatcher.scala
@@ -75,13 +75,12 @@ private[finagle] abstract class GenStreamingSerialServerDispatcher[Req, Rep, In,
         val save = Local.save()
         val dispatched = try {
           Contexts.local.let(RemoteInfo.Upstream.AddressCtx, trans.remoteAddress) {
-            trans.context.peerCertificate match {
-              case None => dispatch(req)
-              case Some(cert) =>
-                Contexts.local.let(Transport.peerCertCtx, cert) {
-                  dispatch(req)
-                }
-            }
+            val peerCertificates = trans.context.sslSessionInfo.peerCertificates
+            if (peerCertificates.isEmpty) dispatch(req)
+            else
+              Contexts.local.let(Transport.peerCertCtx, peerCertificates.head) {
+                dispatch(req)
+              }
           }
         } finally Local.restore(save)