+++ finagle-core/src/main/scala/com/twitter/finagle/dispatch/ServerDispatcher.scala
@@ -69,13 +69,12 @@ abstract class GenSerialServerDispatcher[Req, Rep, In, Out](trans: Transport[In,
       val save = Local.save()
       val dispatched = try {
         Contexts.local.let(RemoteInfo.Upstream.AddressCtx, trans.remoteAddress) {
-          trans.context.peerCertificate match {
-            case None => dispatch(req, eos)
-            case Some(cert) =>
-              Contexts.local.let(Transport.peerCertCtx, cert) {
-                dispatch(req, eos)
-              }
-          }
+          val peerCertificates = trans.context.sslSessionInfo.peerCertificates
+          if (peerCertificates.isEmpty) dispatch(req, eos)
+          else
+            Contexts.local.let(Transport.peerCertCtx, peerCertificates.head) {
+              dispatch(req, eos)
+            }
         }
       } finally Local.restore(save)