+++ finagle-core/src/test/scala/com/twitter/finagle/pushsession/utils/MockChannelHandle.scala
@@ -2,9 +2,9 @@ package com.twitter.finagle.pushsession.utils
 
 import com.twitter.finagle.Status
 import com.twitter.finagle.pushsession.{PushChannelHandle, PushSession}
+import com.twitter.finagle.ssl.session.{NullSslSessionInfo, SslSessionInfo}
 import com.twitter.util.{Future, Promise, Return, Time, Try}
 import java.net.{InetSocketAddress, SocketAddress}
-import java.security.cert.Certificate
 import scala.collection.mutable
 
 class MockChannelHandle[In, Out](var currentSession: PushSession[In, Out])
@@ -69,7 +69,7 @@ class MockChannelHandle[In, Out](var currentSession: PushSession[In, Out])
     pendingWrites += SendAndForgetMany(messages.toVector)
   }
 
-  def peerCertificate: Option[Certificate] = None
+  def sslSessionInfo: SslSessionInfo = NullSslSessionInfo
 
   val remoteAddress: SocketAddress = new InetSocketAddress(8888)