+++ finagle-netty4/src/main/scala/com/twitter/finagle/netty4/transport/ChannelTransportContext.scala
@@ -1,10 +1,10 @@
 package com.twitter.finagle.netty4.transport
 
+import com.twitter.finagle.ssl.session.{NullSslSessionInfo, SslSessionInfo, UsingSslSessionInfo}
 import com.twitter.finagle.transport.TransportContext
 import io.netty.channel.Channel
 import io.netty.handler.ssl.SslHandler
 import java.net.SocketAddress
-import java.security.cert.Certificate
 import java.util.concurrent.Executor
 import scala.util.control.NonFatal
 
@@ -20,15 +20,16 @@ private[finagle] final class ChannelTransportContext(val ch: Channel)
 
   def remoteAddress: SocketAddress = ch.remoteAddress
 
-  val peerCertificate: Option[Certificate] = ch.pipeline.get(classOf[SslHandler]) match {
-    case null => None
-    case handler =>
-      try {
-        handler.engine.getSession.getPeerCertificates.headOption
-      } catch {
-        case NonFatal(_) => None
-      }
-  }
+  val sslSessionInfo: SslSessionInfo =
+    ch.pipeline.get(classOf[SslHandler]) match {
+      case null => NullSslSessionInfo
+      case handler =>
+        try {
+          new UsingSslSessionInfo(handler.engine.getSession)
+        } catch {
+          case NonFatal(_) => NullSslSessionInfo
+        }
+    }
 
   def executor: Executor = ch.eventLoop
 }