+++ finagle-netty4/src/test/scala/com/twitter/finagle/netty4/transport/ChannelTransportContextTest.scala
@@ -1,5 +1,6 @@
 package com.twitter.finagle.netty4.transport
 
+import com.twitter.finagle.ssl.session.NullSslSessionInfo
 import io.netty.channel.embedded.EmbeddedChannel
 import io.netty.handler.ssl.SslHandler
 import java.security.cert.{Certificate, X509Certificate}
@@ -22,23 +23,25 @@ class ChannelTransportContextTest extends FunSuite with MockitoSugar {
     assert(context.remoteAddress == ch.remoteAddress)
   }
 
-  test("peer certificate is empty by default") {
+  test("sslSessionInfo is not used by default") {
     val ch = new EmbeddedChannel()
     val context = new ChannelTransportContext(ch)
-    assert(context.peerCertificate.isEmpty)
+    assert(context.sslSessionInfo == NullSslSessionInfo)
   }
 
-  test("peer certificate is retrieved from an existing SSLEngine") {
+  test("sslSessionInfo is created based on an existing SSLSession") {
     val cert = mock[X509Certificate]
     val certs = Array[Certificate](cert)
     val session = mock[SSLSession]
+    when(session.getId).thenReturn("abcd".getBytes)
+    when(session.getLocalCertificates).thenReturn(Array[Certificate]())
     when(session.getPeerCertificates).thenReturn(certs)
     val engine = mock[SSLEngine]
     when(engine.getSession).thenReturn(session)
     val sslHandler = new SslHandler(engine)
     val ch = new EmbeddedChannel(sslHandler)
     val context = new ChannelTransportContext(ch)
-    assert(context.peerCertificate.nonEmpty)
+    assert(context.sslSessionInfo.peerCertificates.nonEmpty)
   }
 
   test("executor is the channel's event loop") {